Add files via upload
Se actualizan los controladores post y user para corregir cierre de session y se agregan algunos métodos para renderizar nuevas vistas

+++ src/controllers/post.controller.js
@@ -2,6 +2,7 @@
 
 //REQUIRES
 const postService = require('../services/post.service');
+const userService = require('../services/user.service');
 
 //CLASE CONTROLADOR DE POSTS
 class controllerPosts {
@@ -11,6 +12,7 @@ class controllerPosts {
         if(!req.session.userId) {
             return res.redirect('/');
         }
+        const userId = req.session.userId;
         let message = "";
         const title = "ALL POSTS";
         const header = "LISTADO DE TODOS LOS POST";
@@ -20,7 +22,7 @@ class controllerPosts {
                 order: [['title', 'ASC']],
             };
             const posts = await postService.getAllPosts(options);
-            res.status(200).render('posts', { posts, title, header, templateName, message });
+            res.status(200).render('posts', { posts, userId, title, header, templateName, message });
         } catch(error) {
             return res.status(400).send({ message: error.message });
         }
@@ -47,25 +49,43 @@ class controllerPosts {
     }
 
     //CONTROLLER FORMULARIO DE CREACIÓN DE PUBLICACIÓN
-    async formPost(req, res) {
+    async postForm(req, res) {
         if(!req.session.userId) {
             return res.redirect('/');
         }
-        const postBody = req.body;
+        let message = "";
         try {
-            const message = await postService.createPost(postBody);
-            return res.status(200).send({ message });
+            
+            const userId = req.session.userId;
+            const user = await userService.getUserById(userId);
+            const firstName = user.firstName;
+            const lastName = user.lastName;
+            const title = "POST CREATE";
+            const header = "CREACIÓN DE UNA PUBLICACIÓN";
+            const templateName = "postcreate";
+            return res.status(200).render('postcreate', { userId, firstName, lastName, title, header, templateName, message });
         } catch(error) { 
                 return res.status(400).send({ message: error.message });
         }
     }
 
-    //CONTROLLER PROCESAMIENTO DE LA PUBLICACIÓN
-    async processPost(req, res) {
+    //CONTROLLER PROCESAMIENTO DE LA CREACIÓN DE LA PUBLICACIÓN
+    async postProcess(req, res) {
         const postBody = req.body;
+        const title = "ALL POSTS";
+        const header = "LISTADO DE TODOS LOS POST";
+        const templateName = "posts";
+        const userId = req.session.userId;
+        const user = userService.getUserById(userId);
+        const firstName = user.firstName;
+        const lastName = user.lastName;
         try {
             const message = await postService.createPost(postBody);
-            return res.status(200).send({ message });
+            const options = {
+                order: [['title', 'ASC']],
+            };
+            const posts = await postService.getAllPosts(options);
+            res.status(200).render('posts', { posts, userId, firstName, lastName, title, header, templateName, message });
         } catch(error) { 
                 return res.status(400).send({ message: error.message });
         }
@@ -88,7 +108,6 @@ class controllerPosts {
             return res.status(200).render('postedit', { post, title, firstName, lastName, header, templateName, message });
         } catch (error) {
             message = error;
-            console.log(error);
             const title = "FORO ESTUDIANTIL";
             const header = "BIENVENID@, ESTE ES UN ESPACIO PENSADO PARA VOS...";
             const templateName = "index";
@@ -96,19 +115,19 @@ class controllerPosts {
         }
     }
 
-    //CONTROLLER ACTUALIZA DATOS DE UNA PUBLICACIÓN
+    //CONTROLLER EDICION DE DATOS DE UNA PUBLICACIÓN
     async updatePost(req, res) {
         const { guid } = req.params;
         const postBody = req.body;
-        const title = "POST EDIT";
-        const header = "EDICIÓN DE PUBLICACIÓN";
-        const templateName = "postedit";
-        const { post, user } = await postService.getPostById(guid);
-        const firstName = user.firstName;
-        const lastName = user.lastName;
+        const title = "POST";
+        const templateName = "post";
         try {
             const message = await postService.updatePost(guid, postBody);
-            return res.status(200).render('postedit', { post, title, firstName, lastName, header, templateName, message });
+            const { post, user } = await postService.getPostById(guid);
+            const header = `PUBLICACIÓN « ${post.title} »`;
+            const firstName = user.firstName;
+            const lastName = user.lastName;
+            return res.status(200).render('post', { post, title, firstName, lastName, header, templateName, message });
         } catch (error) {
             return res.status(400).render('index',{ message });
         }
@@ -116,10 +135,20 @@ class controllerPosts {
 
     //CONTROLLER ELIMINAR UNA PUBLICACIÓN
     async deletePost(req, res) {
+        if(!req.session.userId) {
+            return res.redirect('/');
+        }
         const { guid } = req.params;
+        const title = "ALL POSTS";
+        const header = "LISTADO DE TODOS LOS POST";
+        const templateName = "posts";
         try {
             const message = await postService.deletePost(guid);
-            return res.status(200).send({ message });
+            const options = {
+                order: [['title', 'ASC']],
+            };
+            const posts = await postService.getAllPosts(options);
+            res.status(200).render('posts', { posts, title, header, templateName, message });
         } catch(error) {
             return res.status(400).send({ message: error.message });
         }
